firmware: Refactor alarms Manager
