+++ firmware/ventilator-controller-stm32/Core/Src/main.cpp
@@ -32,12 +32,13 @@
 #include <functional>
 
 #include "Pufferfish/AlarmsManager.h"
+#include "Pufferfish/Application/Alarms.h"
 #include "Pufferfish/Application/LogEvents.h"
 #include "Pufferfish/Application/States.h"
 #include "Pufferfish/Application/mcu_pb.h"  // Only used for debugging
 #include "Pufferfish/Driver/BreathingCircuit/AlarmLimitsService.h"
 #include "Pufferfish/Driver/BreathingCircuit/AlarmMuteService.h"
-#include "Pufferfish/Driver/BreathingCircuit/Alarms.h"
+#include "Pufferfish/Driver/BreathingCircuit/AlarmsService.h"
 #include "Pufferfish/Driver/BreathingCircuit/ControlLoop.h"
 #include "Pufferfish/Driver/BreathingCircuit/ParametersService.h"
 #include "Pufferfish/Driver/BreathingCircuit/SignalSmoothing.h"
@@ -350,7 +351,7 @@ int interface_test_state = 0;
 int interface_test_millis = 0;
 
 // Alarms
-PF::Driver::BreathingCircuit::AlarmsManager alarms_manager(log_events_manager);
+PF::Application::AlarmsManager alarms_manager(log_events_manager);
 PF::Driver::BreathingCircuit::AlarmsServices breathing_circuit_alarms;
 
 // Breathing Circuit Control
